Fix some event-dispatching tests and move them into a separate file
# Conflicts:
#	compose/ui/ui/src/skikoTest/kotlin/androidx/compose/ui/platform/RenderPhasesTest.kt
